Update airnefcmd.py
Modified the main script to stop displaying the waiting X seconds messages on the console, so that docker won't save them as log entries.

Also modified the code to take out the retry attempt counter, since it isn't really needed if we want this to run continuously.

Lastly modified the Main loop to continue running on successful downloads, and only exit on an error.  This keeps the container from shutting down after a download.